fix: use d2 adapter
