--- src/custom-css/D2UISidebarOverrides.css
@@ -1,7 +1,0 @@
-.left-bar {
-    position: fixed !important;
-    bottom: 0 !important;
-    top: 0 !important;
-    left: 0 !important;
-    margin-top: 3rem !important;
-}
