--- src/index.css
@@ -1,17 +1,0 @@
-html {
-    background: #f3f3f3;
-    font-family: 'Roboto', sans-serif;
-}
-
-body {
-    margin: 0;
-    padding: 0;
-}
-
-h1 {
-    font-family: 'Roboto', sans-serif;
-    font-size: 24px;
-    font-weight: 300;
-    letter-spacing: 1.2px;
-    color: rgba(0, 0, 0, 0.87);
-}
